+++ src/api/routers/professors.py
@@ -11,22 +11,20 @@ from fastapi_cache.coder import PickleCoder
 from constants import Constants
 import os
 
-class ProfessorRoutes:
-    def __init__(self):
-        db_conn = connection.db
-        professor_info = All_professors.Professor(db_conn, FastAPICache)
-        api_key_header = APIKeyHeader(name="X-API-Key")
+db_conn = connection.db
+professor_info = All_professors.Professor(db_conn, FastAPICache)
+api_key_header = APIKeyHeader(name="X-API-Key")
 
-    def get_api_key(api_key_header: str = Security(api_key_header)) -> str:
-        if api_key_header in os.environ.get("API_SIGN_KEY"):
-            return api_key_header
-        raise HTTPException(
-            status_code=Request.status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid or missing API Key",
-        )
+def get_api_key(api_key_header: str = Security(api_key_header)) -> str:
+    if api_key_header in os.environ.get("API_SIGN_KEY"):
+        return api_key_header
+    raise HTTPException(
+        status_code=Request.status.HTTP_401_UNAUTHORIZED,
+        detail="Invalid or missing API Key",
+    )
 
 router = APIRouter(
-    prefix='api/professors',
+    prefix='/professors',
     tags=['professors'],
     responses={404: {'description': 'Not found'}}
 )
